face_recognition_module/requirements.txt
opencv-python
torch
faiss-cpu
facenet-pytorch
numpy
